CHORE: Replace SQL module with a more up-to-date one
The MySQL module used (mysql-connector-python-rf) is unofficial and used to be de facto, as the official version had major bugs. The official module from Oracle (mysql-connector-python) has since been updated, and mysql-connector-python-rf has not seen updates in years. This commit replaces the unofficial with the official one, and has been demonstrated to be working via unit tests.